Address more PR feedback
