+++ mountpoint-s3-fs/src/cli.rs
@@ -61,16 +61,16 @@ pub struct ContextParams {
             .multiple(true),
     )
 )]
-#[command(after_help = concat!(
-    "\nAlternative fstab structure:\n",
+#[command(after_help = if cfg!(feature = "fstab") {concat!(
+    "\nAlternative fstab style:\n",
     "  mount-s3 <BUCKET> <DIRECTORY> -o <OPTIONS>\n\n",
     "Arguments:\n",
     "  <BUCKET_NAME>\n          Name of bucket to mount, with s3:// URIs supported\n",
     "  <DIRECTORY>\n          Location to mount bucket at\n",
     "  <OPTIONS>\n",
     "          fstab style options. Comma separated list of CLI options, with backslash escapes for commas, backslashes, and double quotes.\n",
-    "          No need to use `--` to prefix arguments."
-))]
+    "          Use of `--` to prefix arguments is optional."
+)} else {""})]
 pub struct CliArgs {
     #[clap(help = "Name of bucket to mount", value_parser = parse_bucket_name)]
     pub bucket_name: String,
