+++ mountpoint-s3/Cargo.toml
@@ -46,8 +46,8 @@ built = { version = "0.7.5", features = ["git2"] }
 # Unreleased feature flags
 block_size = ["mountpoint-s3-fs/block_size"]
 event_log = ["mountpoint-s3-fs/event_log"]
+fstab = ["mountpoint-s3-fs/fstab"]
 mem_limiter = ["mountpoint-s3-fs/mem_limiter"]
-fstab = []
 # Features for choosing tests
 s3_tests = ["mountpoint-s3-fs/s3_tests"]
 fuse_tests = ["mountpoint-s3-fs/fuse_tests"]
