+++ mountpoint-s3/src/main.rs
@@ -1,8 +1,8 @@
 use mountpoint_s3::build_info;
-use mountpoint_s3::cli::get_cli_args;
+use mountpoint_s3::try_parse_cli_args;
 
 fn main() -> anyhow::Result<()> {
-    let cli_args = get_cli_args()?;
+    let cli_args = try_parse_cli_args()?;
 
     let context = mountpoint_s3_fs::cli::ContextParams {
         full_version: build_info::FULL_VERSION.to_string(),
